PEBackend/requirements.txt
7==0.0.0
asgiref==3.2.3
bson==0.5.8
cffi==1.14.0
cryptography==2.8
Django==2.2.10
django-extensions==2.2.9
django-cors-headers==3.2.1
djangorestframework==3.11.0
djongo==1.3.2
PyJWT==1.7.1
pycparser==2.20
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.21.0
selenium==3.141.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.24.2